requirements.txt
Pillow
pydantic==1.10
requests
